chore: introduce feature flag to swap metric explore modal v1 -> v2
