+++ packages/common/src/types/featureFlags.ts
@@ -84,6 +84,12 @@ export enum FeatureFlags {
      * Use screenshot ready indicator for scheduler screenshots
      */
     ScreenshotReadyIndicator = 'screenshot-ready-indicator',
+
+    /**
+     * Enable echarts visualization in Metrics Catalog explore modal (V2)
+     * When enabled, uses VisualizationProvider + echarts instead of recharts
+     */
+    MetricsCatalogEchartsVisualization = 'metrics-catalog-echarts-visualization',
 }
 
 export type FeatureFlag = {
